squashme: minor fix

+++ cleanup-renku-ci-deployments/entrypoint.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 if test -z "$RENKUBOT_KUBECONFIG" ; then
     echo 'Please specify a kubeconfig that can be used for the helm and kubectl commands.'
@@ -16,31 +16,31 @@ HELM_RELEASE_REGEX="${HELM_RELEASE_REGEX:=".*"}"
 K8S_CI_NAMESPACE_REGEX=".+-ci-.+|^ci-.+"
 MAX_AGE_SECONDS=${MAX_AGE_SECONDS:=604800}
 GITLAB_URL="https://dev.renku.ch/gitlab"
-KUBECONFIG_LINES=$(wc -l $KUBECONFIG)
 DELETE_NAMESPACE=${DELETE_NAMESPACE:="false"}
 
 echo "$RENKUBOT_KUBECONFIG" > "$KUBECONFIG" && chmod 400 "$KUBECONFIG"
 echo "Kubeconfig is at $KUBECONFIG."
+KUBECONFIG_LINES=$(wc -l $KUBECONFIG)
 echo "Kubeconfig is $KUBECONFIG_LINES long."
 echo "Looking for CI releases with regex $HELM_RELEASE_REGEX."
-echo "Looking in namespaces with regex $HELM_RELEASE_REGEX."
+echo "Looking in namespaces with regex $K8S_CI_NAMESPACE_REGEX."
 echo "Age threshold for deletion is $MAX_AGE_SECONDS seconds."
 echo "Delete namespace: $DELETE_NAMESPACE"
 
 NOW=$(date +%s)
 NAMESPACES=$(kubectl get namespaces -o json | jq -cr ".items | map(select(.metadata.name | test(\"$K8S_CI_NAMESPACE_REGEX\"))) | .[].metadata.name")
-for NAMESPACE in NAMESPACES 
+for NAMESPACE in $NAMESPACES 
 do
     RELEASES=$(helm list -n $NAMESPACE --all -f "$HELM_CI_RELEASE_REGEX" -o json | jq -cr "map(select(.name | test(\"$HELM_RELEASE_REGEX\"))) | .[].name")
     for RELEASE in $RELEASES
     do
-        if [ ! -z $RELEASE ] && [ ! -z $NAMESPACE ]
+        if [[ ! -z $RELEASE ]] && [[ ! -z $NAMESPACE ]]
         then
             echo "Checking release $RELEASE in namespace $NAMESPACE."
             # extract last deployed date and convert to unix epoch
             LAST_DEPLOYED_AT=$(helm -n $NAMESPACE history $RELEASE -o json | jq -r 'last | .updated | sub("\\.[0-9]{6}.*$"; "Z") | fromdateiso8601')
             AGE_SECONDS=$(expr $NOW - $LAST_DEPLOYED_AT)
-            if [ $AGE_SECONDS -ge $MAX_AGE_SECONDS ] || [ $MAX_AGE_SECONDS -le 0 ]
+            if [[ $AGE_SECONDS -ge $MAX_AGE_SECONDS ]] || [[ $MAX_AGE_SECONDS -le 0 ]]
             then
                 # remove any jupyterservers - they have finalizers that prevent the namespces to be deleted
                 SERVERS=$(kubectl -n $NAMESPACE get jupyterservers -o json | jq -r '.items | .[].metadata.name') 
@@ -75,9 +75,12 @@ do
             echo "Release $RELEASE and/or $NAMEPSACE are empty."
         fi
     done
-    if [ "$DELETE_NAMESPACE" = "true" ] && [ ! -z $NAMESPACE ] && [ -z $RELEASES ]
+    if [[ "$DELETE_NAMESPACE" = "true" ]] && [[ ! -z $NAMESPACE ]] && [[ -z $RELEASES ]] && [[ $NAMESPACE =~ $HELM_RELEASE_REGEX ]]
     then
-        # remove the namespace if there are no releases in it
+        # Remove the namespace if there are no releases in it 
+        # and the namespace name matches the name of the release - all CI deployments follow this pattern.
+        # This addresses the case when a CI deployment does not persist but its namespace persists,
+        # so when the PR is closed only the namespace in k8s remains and should be cleaned up.
         echo "Deleting namespace $NAMESPACE"
         kubectl delete namespace $NAMESPACE --wait
     fi
